+++ egui_demo_lib/src/apps/demo/plot_demo.rs
@@ -75,8 +75,7 @@ impl LineDemo {
                     .on_hover_text("Always keep the viewport square.");
                 ui.checkbox(proportional, "Proportional data axes")
                     .on_hover_text("Tick are the same size on both axes.");
-            });
-            ui.vertical(|ui| {
+
                 ComboBox::from_label("Line style")
                     .selected_text(line_style.to_string())
                     .show_ui(ui, |ui| {