+++ CHANGELOG.md
@@ -8,6 +8,12 @@ NOTE: [`eframe`](eframe/CHANGELOG.md), [`egui_web`](egui_web/CHANGELOG.md) and [
 ## Unreleased
 
 
+## 0.14.2 - 2021-08-28 - Window resize fix
+
+### Fixed 🐛
+* Fix window resize bug introduced in `0.14.1`.
+
+
 ## 0.14.1 - 2021-08-28 - Layout bug fixes
 
 ### Added ⭐